Removed the computation and analysis of subject-specific Cohen's d: This analysis would lead to a double testing of the FAB based on the same value (RT). Therefore, we remove it and instead check the association between RADS and FAB as well as between ASRS and FAB with correlations.
Additionally, we adapted the model that was supposed to check for FAB regardless of saccades. Previously, we had computed median RTs ignoring trials on which saccades were produced. However, specifically in the BOTH group, there are some participants with saccades on more than half of the trials, leading to immense data loss. Therefore, we include a predictor n.sac that assesses how the number of saccades influences the RT.